fix: move test to more appropriate context
